requirements.txt
fastapi
uvicorn[standard]
aiogram==3.*
python-dotenv
SQLAlchemy
psycopg[binary]
python-dotenv
httpx

